end of glish v2

+++ src/Texture.cpp
@@ -3,6 +3,7 @@
 //
 
 
+#include <SDL2/SDL.h>
 #include <SDL2/SDL_image.h>
 #include <glish/Texture.hpp>
 #include <iostream>
@@ -144,7 +145,7 @@ namespace glish {
     Texture Texture::makeTextureCubeMap(const std::array<std::string, 6> &pathImage) {
 	    Texture texture(GL_TEXTURE_CUBE_MAP);
 
-	    for(size_t i = 0; i < pathImage.size() ; i++){
+	    for(unsigned i = 0; i < pathImage.size() ; i++){
 		    texture.loadTexture(pathImage[i], GL_TEXTURE_CUBE_MAP_POSITIVE_X +i);
 	    }
